refactor(cli): reorganize action handlers into a dedicated subpackage
