+++ cli/src/main/kotlin/com/github/adriianh/cli/command/player/SearchCommand.kt
@@ -1,5 +1,6 @@
 package com.github.adriianh.cli.command.player
 
+import com.github.adriianh.cli.command.player.handler.SearchActionHandler
 import com.github.adriianh.cli.config.Messages
 import com.github.adriianh.cli.config.configDir
 import com.github.adriianh.cli.config.resolveEnv
